p4i integration Integrate 116144 from 3 to main: Remove warnings.
[git-p4: depot-paths = "//depot/qt/main/": change = 116145]

+++ src/canvas/qcanvas.cpp
@@ -4433,10 +4433,10 @@ void QCanvasLine::moveBy(double dx, double dy)
 {
     if (dx || dy) {
 	removeFromChunks();
-	x1 += dx;
-	y1 += dy;
-	x2 += dx;
-	y2 += dy;
+	x1 += (int)dx;
+	y1 += (int)dy;
+	x2 += (int)dx;
+	y2 += (int)dy;
 	addToChunks();
     }
 }
